pyproject.toml
[tool.poetry]
name = "gamearmy-telegram-bot"
version = "0.2.0"
description = "Telegram bot for monitoring game servers on gamearmy.ru"
authors = ["dealenx <dealenx@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
default = "gamearmy_telegram_bot.main:run"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
